requirements.txt
requests
matplotlib
numpy
azure-storage-blob